+++ src/core/perf/config/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2022-??-??)
+
+#### :rocket: New Feature
+
+* Added ability to use include, exclude filters.
+
+
 ## v3.61.0 (2021-10-20)
 
 #### :rocket: New Feature